i386: hvf: Use void * for register pointers
target_ulong has different size for qemu-system-i386 and
qemu-system-x86_64: 32 and 64 bits, respectively. The type cannot be
used to store pointers to registers in the host memory. An attempt to
cast a register pointer to target_ulong can result in truncated pointer
and SIGSEGV in qemu-system-i386.

write_val_ext and read_val_ext are mixed functions that accept either
host pointers or guest addresses.

Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>