i386: hvf: Dump vmcs fields on hv_vcpu_run failure
hvf_vcpu_exec asserts that hv_vcpu_run succeeds and aborts the program
otherwise. The return value of hv_vcpu_run is printed to stderr
immediately before the abort but it doesn't provide enough details about
the actual cause of vmexit or vmentry failure.

To assist in debugging, Intel provides instruction error, exit reason
and exit qualification fields in VMCS. They are set by
vmlaunch/vmresume.  It will be helpful to print the fields before the
assert and in the case of EPT violation exit reason, it's worth to dump
EPTP, Guest Physical Address and Guest Linear Address.

Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>