hmp: Add 'info accel' command
The command is made after 'info kvm' and aims to replace it as more
generic one.

If used without parameters, the command can used to get current
accelerator. Otherwise, it may be used to determine if an accelerator is
available. Here's an example if a VM with hvf accel is started:

  (qemu) info accel
  hvf support: enabled
  (qemu) info accel kvm
  kvm support: not compiled
  (qemu) info accel tcg
  tcg support: disabled

Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>